requirements.txt
# Library dependencies for the python code.  You need to install these with
# `pip install -r requirements.txt` before you can run this.

# Classic machine learning packages.
numpy<1.19.0,>=1.16.0
pandas >= 0.24.0
scikit-learn >= 0.21.0
hyperopt >= 0.2.5
neptune-client >= 0.4.125

# Deep learning packages.
torch >= 1.5.0
pytorch_lightning >= 1.0.5
tensorflow >= 2.0.0
tensorboardX >= 1.9
transformers == 3.1.0

# Infrasturcture: packages for configuration and database management.
python-dateutil >= 2.8.1
jsonnet >= 0.10.0
sqlalchemy >= 1.2.0
protobuf >= 3.10.0
flask >= 1.1.1

# Natural language processing packages.
nltk >= 3.2.5
pyonmttok >= 1.22.1
allennlp >= 1.1.0
fasttext >= 0.9.1
slovnet == 0.4.0
pymorphy2[fast] == 0.9.1

# Packages for visualizations.
jupyter >= 1.0.0
seaborn >= 0.9.0
ipymarkup >= 0.5.0

# Other packages
pyinstrument == 3.2.0
hnswlib == 0.4.0
